% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrcpppkg-package.R
\docType{package}
\name{testrcpppkg-package}
\alias{testrcpppkg-package}
\alias{testrcpppkgpackage}
\title{Rcpp Test Package.}
\description{
Smorgasbord for remote sensing functions
}
\details{
The package provides a variety of functions which are useful for
handling, manipulating and visualizing remote sensing data.
}
\keyword{package}
